+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - DF-996 Fixed API Docs to show token refresh endpoints
+- DF-982 Fixed GET over POST on system resources ignoring some parameters in payload
 
 ## [0.9.1] - 2017-04-25
 ### Fixed